extract UtxoRepositoryImpl class to LedgerLibUtxoRespositoryImpl
