+++ libs/persistence/build.gradle
@@ -14,10 +14,10 @@ dependencies {
 
     implementation platform("net.corda:corda-api:$cordaApiVersion")
 
-//    implementation project(':libs:crypto:cipher-suite')
-//    implementation project(':libs:crypto:crypto-core')
 //    implementation project(':libs:flows:external-event-responses')
 //    implementation project(':libs:flows:flow-utils')
+    implementation project(':libs:crypto:crypto-core')
+    implementation project(':libs:crypto:cipher-suite')
     implementation project(':libs:ledger:ledger-common-data')
 //    implementation project(':libs:ledger:ledger-consensual-data')
     implementation project(':libs:ledger:ledger-utxo-data')
